Fix typing hints for timeouts: change from int to float
Checking both Paramiko and Telnetlib, all those parameters are actually float.

Only self.timeout seems to be used as an int in Netmiko itself, so it's
left unchanged.

The keepalive parameters was wrongly typed as an integer in typeshed, this
has been fixed in https://github.com/python/typeshed/pull/14596